app/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
bitarray==1.2.2
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cycler==0.11.0
cytoolz==0.11.2
debugpy==1.6.0
decorator==5.1.1
dnspython==2.2.1
entrypoints==0.4
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
etherscan-python==2.1.0
executing==0.8.3
fastapi==0.75.1
fonttools==4.32.0
frozenlist==1.3.0
h11==0.12.0
hexbytes==0.2.2
httpcore==0.14.7
httpx==0.22.0
idna==3.3
ipfshttpclient==0.8.0a2
ipykernel==6.13.0
ipython==8.2.0
jedi==0.18.1
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.2.2
jupyter-core==4.9.2
kiwisolver==1.4.2
lru-dict==1.1.7
matplotlib==3.5.1
matplotlib-inline==0.1.3
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
nest-asyncio==1.5.5
netaddr==0.8.0
numpy==1.22.3
packaging==21.3
pandas==1.4.2
parsimonious==0.8.1
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.1
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.0
pure-eval==0.2.2
pycryptodome==3.14.1
pydantic==1.9.0
Pygments==2.11.2
pymongo==4.1.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pywin32==303
pyzmq==22.3.0
ratelimit==2.2.1
requests==2.27.1
rfc3986==1.5.0
rlp==2.0.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
stack-data==0.2.0
starlette==0.17.1
threadpoolctl==3.1.0
tomli==2.0.1
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
varint==1.0.2
wcwidth==0.2.5
web3==5.28.0
websockets==9.1
yarl==1.7.2
